代码优化

+++ core/data/src/main/kotlin/com/zhixue/lite/core/data/repository/UpdateRepository.kt
@@ -2,5 +2,5 @@ package com.zhixue.lite.core.data.repository
 
 interface UpdateRepository {
 
-    suspend fun getNewVersionCode(): Int
+    suspend fun getNewVersionCode(): Long
 }
