+++ core/data/src/main/kotlin/com/zhixue/lite/core/data/repository/UpdateRepositoryImpl.kt
@@ -7,7 +7,7 @@ class UpdateRepositoryImpl @Inject constructor(
     private val networkDataSource: ApiNetworkDataSource
 ) : UpdateRepository {
 
-    override suspend fun getNewVersionCode(): Int {
+    override suspend fun getNewVersionCode(): Long {
         return networkDataSource.getNewVersionCode().newVersionCode
     }
 }
