+++ feature/profile/src/main/kotlin/com/zhixue/lite/feature/profile/ProfileViewModel.kt
@@ -63,7 +63,7 @@ class ProfileViewModel @Inject constructor(
     }
 
     fun checkUpdate(
-        versionCode: Int,
+        versionCode: Long,
         onNewVersionAvailable: () -> Unit
     ) {
         viewModelScope.launch {
